+++ test/e2e/storage/empty_dir_wrapper.go
@@ -211,7 +211,7 @@ var _ = utils.SIGDescribe("EmptyDir wrapper volumes", func() {
 func createGitServer(ctx context.Context, f *framework.Framework) (gitURL string, gitRepo string, cleanup func()) {
 	var err error
 	gitServerPodName := "git-server-" + string(uuid.NewUUID())
-	containerPort := 8000
+	containerPort := int32(8000)
 
 	labels := map[string]string{"name": gitServerPodName}
